Refactor and merge random search utils into random_search module.
This aggregates the random search utility classes and functions into
the random_search module. This effectively deprecates the
compiler_gym.random_replay and compiler_gym.util.logs modules.
