+++ compiler_gym/bin/random_replay.py
@@ -17,7 +17,7 @@ from pathlib import Path
 from absl import app, flags
 
 import compiler_gym.util.flags.output_dir  # noqa Flag definition.
-from compiler_gym.random_replay import replay_actions_from_logs
+from compiler_gym.random_search import replay_actions_from_logs
 from compiler_gym.util import logs
 from compiler_gym.util.flags.benchmark_from_flags import benchmark_from_flags
 from compiler_gym.util.flags.env_from_flags import env_from_flags
